requirements.txt
# Core dependencies
browser-use>=0.1.29
python-dotenv>=1.0.1
flask>=3.0.0

# LLM dependencies
langchain-openai>=0.0.3
pydantic>=2.0.0

# Browser automation dependencies
playwright>=1.49.0

# Content extraction dependencies
lxml[html_clean]>=5.3.0
MainContentExtractor>=0.0.4 